requirements.txt
#utility
python-dotenv
pandas
numpy
packaging  #Version parsing and comparison
typer   #to create an CLI application    Ex: python pdf_assistance.py <variable_name> 
torch

#llm
groq
openai
google-generativeai

#genai-framework
langchain
phidata 

langchain_community
langchainhub
langchain-google-genai
langchain-groq
langchain-nomic
langchain-text-splitters
langchain_mistralai
langchain_huggingface
langgraph

#tools
duckduckgo-search
yfinance

tavily-python
wikipedia
google-search-results
youtube_search
yfinance

#deployment
fastapi
uvicorn
streamlit

#db
sqlalchemy
pgvector
psycopg[binary]

chromadb
faiss-cpu

#reader
pypdf

#embediings
sentence-transformers

#chunking & tokenization
tiktoken
sentence_transformers